refactor wear and takeoff
