+++ tests/strategies/test_fight_ranged.py
@@ -33,7 +33,7 @@ class TestCrossLavaRive:
             argv=[
                 f"--env={env}",
                 f"--seed={seed}",
-                # "--no-render",
+                "--no-render",
                 "--autopickup=True",
                 "--code_wrapper=False",
             ]
